Completely refactored component names
In a previous commit the refactoring of component names have not been
done completely : this commit finished the work.

The naming convention is : prefix the component name with C when it is a
component, a P for a page, a R for a route...

This has been made to differentiate the component / pages / routes with
data models.